Consolidate integration test suites
Right now we have full test suites for running just a couple tests. The
overhead of each suite is fairly high, so it makes sense to consolidate
these a bit.

This removes the istioctl and framework targets.
