+++ tests/integration/galley/analyze_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package versionanalyze
+package galley
 
 import (
 	"strings"
