+++ tests/integration/galley/testdata/reviews-svc.yaml
