+++ tests/integration/galley/testdata/servicerolebinding.yaml
