+++ tests/integration/security/webhook/config/galley-webhook.yaml
